Changing Player Class
Added in a way to ensure the player's stats do not go below 8 for
character generation. It will continue to reroll the stat until it is
above 8. Intelligence needs to be changed in any instance that it is
used to Vitality. I did not change it on the player class yet as I did
not want to break half the program. I can change it myself, but if
someone wants to just go through and do that they're more than welcome.