Added in the Combat Module
I added in the combat module itself. It still needs to be connected to
the GUI somehow and maybe it's not entirely perfect. But we can discuss
that in class or something. I added a few more identifiers to the items
in the item class so that combat would have an easier time using the
items. Furthermore I added in the calculation method for the player's
defense. That will need refreshing every time they swap armor or put a
shield on, so hopefully that's possible. There are a few other things
that need to be added to combat but I'd rather add them after I make
sure what's already done is okay as it took a fairly substantial chunk
of time to do.